Seeded DB data and added scafolding controllers for DB entities

+++ Models/Written By.cs
@@ -8,7 +8,7 @@ namespace Project.Models
 {
     public class Written_By
     {
-        public int ISBN { get; set; }
+        public long ISBN { get; set; }
         public int AuthorID { get; set; }
 
         public Book Book { get; set; }
